+++ example_project/settings.py
@@ -1,12 +1,6 @@
 import os
 
-ADMINS = (
-    # ('Your Name', 'your_email@domain.com'),
-)
-
 DEBUG = True
-
-MANAGERS = ADMINS
 
 DATABASES = {
     'default': {
@@ -20,13 +14,9 @@ DATABASES = {
 }
 
 TIME_ZONE = 'UTC'
-
 LANGUAGE_CODE = 'en'
-
 USE_I18N = True
-
 USE_L10N = True
-
 USE_TZ = True
 
 LOCALE_PATHS = [